+++ tests/_tplcache/sub_tpl.js
@@ -102,14 +102,14 @@
                             }
                             
                                          
-                            __p__ += '">' + "\n" + '            <a href="/' + ( __instance__.data['user']["name"] ) + '">' + ( __instance__.data['user']['name'] ) + ( __instance__.data['user']['text'] ) + ' ' + ( Contemplate.n(__instance__.data['i']) + Contemplate.n(__instance__.data['j']) ) + '</a>: <strong>' + ( __instance__.data['user']["text"] ) + '</strong>' + "\n" + '        </div>' + "\n" + '        ';            
-                            if (  Contemplate.has_key(__instance__.data['user'], "key1")  )
+                            __p__ += '">' + "\n" + '            <a href="/' + ( __instance__.data['user']["name"] ) + '">' + ( __instance__.data['user']['name'] ) + '' + ( __instance__.data['user']['text'] ) + ' ' + ( Contemplate.n(__instance__.data['i']) + Contemplate.n(__instance__.data['j']) ) + '</a>: <strong>' + ( __instance__.data['user']["text"] ) + '</strong>' + "\n" + '        </div>' + "\n" + '        ';            
+                            if (  Contemplate.haskey(__instance__.data['user'], "key1")  )
                             {
                             
                                              
                                 __p__ += '' + "\n" + '            <div> User has key &quot;key1&quot; </div>' + "\n" + '        ';            
                             }
-                            else if (  Contemplate.has_key(__instance__.data['user'], "key", "key1")  )
+                            else if (  Contemplate.haskey(__instance__.data['user'], "key", "key1")  )
                             {